Move handlers into ./handlers
