+++ gateway/Dockerfile.build
@@ -6,14 +6,15 @@ RUN go get -d github.com/docker/docker/api/types \
    && go get -d github.com/docker/docker/client \
    && go get github.com/gorilla/mux \
    && go get github.com/prometheus/client_golang/prometheus
-# RUN go get -d github.com/prometheus/client_model/go
+RUN go get -d github.com/Sirupsen/logrus
 
 WORKDIR /go/src/github.com/alexellis/faas/gateway
 
 COPY metrics	metrics
 COPY requests	requests
-COPY tests	    tests
+COPY tests	tests
+COPY handlers	handlers
+
 COPY server.go	.
-COPY proxy.go	.
 
 RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o app .
