+++ gateway/handlers/proxy.go
@@ -1,4 +1,4 @@
-package main
+package handlers
 
 import (
 	"bytes"
@@ -6,12 +6,12 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"net/http"
 	"strconv"
 	"strings"
 	"time"
 
+	"github.com/Sirupsen/logrus"
 	"github.com/alexellis/faas/gateway/metrics"
 	"github.com/alexellis/faas/gateway/requests"
 	"github.com/docker/docker/api/types"
@@ -20,27 +20,26 @@ import (
 	"github.com/gorilla/mux"
 )
 
-// MakeProxy creates a proxy for HTTP web requests which can be routed to a function.
-func MakeProxy(metrics metrics.MetricOptions, wildcard bool, c *client.Client) http.HandlerFunc {
+// makeProxy creates a proxy for HTTP web requests which can be routed to a function.
+func MakeProxy(metrics metrics.MetricOptions, wildcard bool, c *client.Client, logger *logrus.Logger) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		metrics.GatewayRequestsTotal.Inc()
 
 		if r.Method == "POST" {
-			log.Println(r.Header)
+			logger.Infoln(r.Header)
 			header := r.Header["X-Function"]
-			log.Println(header)
-			// fmt.Println(wildcard)
+			logger.Infoln(header)
 
 			if wildcard == true {
 				vars := mux.Vars(r)
 				name := vars["name"]
 				fmt.Println("invoke by name")
-				lookupInvoke(w, r, metrics, name, c)
+				lookupInvoke(w, r, metrics, name, c, logger)
 			} else if len(header) > 0 {
-				lookupInvoke(w, r, metrics, header[0], c)
+				lookupInvoke(w, r, metrics, header[0], c, logger)
 			} else {
 				requestBody, _ := ioutil.ReadAll(r.Body)
-				alexaService := isAlexa(requestBody)
+				alexaService := IsAlexa(requestBody)
 				fmt.Println(alexaService)
 
 				if len(alexaService.Session.SessionId) > 0 &&
@@ -50,7 +49,7 @@ func MakeProxy(metrics metrics.MetricOptions, wildcard bool, c *client.Client) h
 					fmt.Println("Alexa SDK request found")
 					fmt.Printf("SessionId=%s, Intent=%s, AppId=%s\n", alexaService.Session.SessionId, alexaService.Request.Intent.Name, alexaService.Session.Application.ApplicationId)
 
-					invokeService(w, r, metrics, alexaService.Request.Intent.Name, requestBody)
+					invokeService(w, r, metrics, alexaService.Request.Intent.Name, requestBody, logger)
 				} else {
 					w.WriteHeader(http.StatusBadRequest)
 					w.Write([]byte("Provide an x-function header or a valid Alexa SDK request."))
@@ -60,10 +59,10 @@ func MakeProxy(metrics metrics.MetricOptions, wildcard bool, c *client.Client) h
 	}
 }
 
-func isAlexa(requestBody []byte) requests.AlexaRequestBody {
+func IsAlexa(requestBody []byte) requests.AlexaRequestBody {
 	body := requests.AlexaRequestBody{}
 	buf := bytes.NewBuffer(requestBody)
-	fmt.Println(buf)
+	// fmt.Println(buf)
 	str := buf.String()
 	parts := strings.Split(str, "sessionId")
 	if len(parts) > 1 {
@@ -72,18 +71,18 @@ func isAlexa(requestBody []byte) requests.AlexaRequestBody {
 	return body
 }
 
-func lookupInvoke(w http.ResponseWriter, r *http.Request, metrics metrics.MetricOptions, name string, c *client.Client) {
+func lookupInvoke(w http.ResponseWriter, r *http.Request, metrics metrics.MetricOptions, name string, c *client.Client, logger *logrus.Logger) {
 	exists, err := lookupSwarmService(name, c)
 	if err != nil || exists == false {
 		if err != nil {
-			log.Fatalln(err)
+			logger.Fatalln(err)
 		}
 		w.WriteHeader(http.StatusInternalServerError)
 		w.Write([]byte("Error resolving service."))
 	}
 	if exists == true {
 		requestBody, _ := ioutil.ReadAll(r.Body)
-		invokeService(w, r, metrics, name, requestBody)
+		invokeService(w, r, metrics, name, requestBody, logger)
 	}
 }
 
@@ -96,7 +95,7 @@ func lookupSwarmService(serviceName string, c *client.Client) (bool, error) {
 	return len(services) > 0, err
 }
 
-func invokeService(w http.ResponseWriter, r *http.Request, metrics metrics.MetricOptions, service string, requestBody []byte) {
+func invokeService(w http.ResponseWriter, r *http.Request, metrics metrics.MetricOptions, service string, requestBody []byte, logger *logrus.Logger) {
 	metrics.GatewayFunctionInvocation.WithLabelValues(service).Add(1)
 
 	stamp := strconv.FormatInt(time.Now().Unix(), 10)
@@ -105,9 +104,10 @@ func invokeService(w http.ResponseWriter, r *http.Request, metrics metrics.Metri
 	buf := bytes.NewBuffer(requestBody)
 	url := "http://" + service + ":" + strconv.Itoa(8080) + "/"
 	fmt.Printf("[%s] Forwarding request to: %s\n", stamp, url)
+
 	response, err := http.Post(url, "text/plain", buf)
 	if err != nil {
-		log.Println(err)
+		logger.Infoln(err)
 		w.WriteHeader(500)
 		buf := bytes.NewBufferString("Can't reach service: " + service)
 		w.Write(buf.Bytes())
